Make aggregate pushdown optimization generic
This PR moves the aggregate pushdown optimization to a separate module,
with the intention to move this code to
[database-connector](https://github.com/duckdb/database-connector) repo.
