+++ src/CMakeLists.txt
@@ -19,8 +19,11 @@ add_library(
 
 add_library(
   mysql_ext_dbconn OBJECT
-  dbconn/query/query_writer.cpp dbconn/table_scan/filter_pushdown.cpp
+  dbconn/query/query_writer.cpp
+  dbconn/table_scan/filter_pushdown.cpp
   dbconn/table_scan/filter_util.cpp
+  dbconn/optimizer/aggregate_optimizer.cpp
+  dbconn/optimizer/optimizer_util.cpp
   dbconn/optimizer/order_by_and_limit_optimizer.cpp)
 
 set(ALL_OBJECT_FILES
